type safety rule for #define

+++ srcLB/LBDefinitions.h
@@ -22,6 +22,7 @@ static const int LATTICEVELOCITIES[19][3] = {
 #define w3 1.0/36.0   /* if ||c_i|| = sqrt(2) */
 
 // Lattice weights
+
 static const double LATTICEWEIGHTS[19] = {
     w3, w3, w2, w3, w3,
     w3, w2, w3, w2, w1,
@@ -36,5 +37,5 @@ static const double LATTICEWEIGHTS[19] = {
 // Speed of sound
 // TODO: (VS) The value of 1.0/sqrt(3.0) as of now,
 // has been obtained using Julia Arbitraty Precision Arithmetic
-#define C_S 0.57735026918962576
+#define C_S ((double)0.57735026918962576) // double cast for type safety
 #endif
